+++ scripts/train/train.py
@@ -70,48 +70,6 @@ def build_composer_model(model_cfg, tokenizer):
     return COMPOSER_MODEL_REGISTRY[model_cfg.name](model_cfg, tokenizer)
 
 
-def build_composer_peft_model(
-        model_cfg: DictConfig, lora_cfg: DictConfig,
-        tokenizer: PreTrainedTokenizer) -> ComposerHFCausalLM:
-    try:
-        from peft import LoraConfig, get_peft_model
-    except ImportError as e:
-        raise ImportError(
-            'Error importing from peft. Please verify that peft and peft utils '
-            'are installed by running `pip install -e .[peft]` from `llm-foundry/`.'
-            f'Error encountered: {e}')
-
-    # 1) loads a hf model, 2) adds peft modules, 3) wraps it in a ComposerHFCausalLM.
-    print('Building Lora config...')
-    lora_cfg = LoraConfig(**lora_cfg.args)
-
-    print('Building model from HuggingFace checkpoint...')
-    model = MPTForCausalLM.from_pretrained(
-        cfg.model.pretrained_model_name_or_path, trust_remote_code=True)
-    print('Model built!')
-
-    print('Adding Lora modules...')
-    model = get_peft_model(model, lora_cfg)
-    print('Lora modules added!')
-
-    model = ComposerHFCausalLM(model, tokenizer)
-
-    return model
-
-
-def print_trainable_parameters(model) -> None:
-    # Prints the number of trainable parameters in the model.
-    trainable_params = 0
-    all_param = 0
-    for _, param in model.named_parameters():
-        all_param += param.numel()
-        if param.requires_grad:
-            trainable_params += param.numel()
-    print(
-        f'trainable params: {trainable_params} || all params: {all_param} || trainable%: {100 * trainable_params / all_param}'
-    )
-
-
 def build_dataloader(cfg, tokenizer, device_batch_size):
     if cfg.name == 'text':
         return build_text_dataloader(
@@ -205,14 +163,7 @@ def main(cfg):
 
     # Build Model
     print('Initializing model...')
-    with init_context:
-        if cfg.get('lora',
-                   None) is not None:  # frozen model + trainable lora modules
-            model: ComposerHFCausalLM = build_composer_peft_model(
-                cfg.model, cfg.lora, tokenizer)
-            print_trainable_parameters(model)  # should not be 100%
-        else:  # standard model
-            model = build_composer_model(cfg.model, tokenizer)
+    model = build_composer_model(cfg.model, tokenizer)
     cfg.n_params = sum(p.numel() for p in model.parameters())
     print(f'{cfg.n_params=:.2e}')
